+++ nix/consumer-cache-hash.nix
@@ -1,1 +1,1 @@
-"sha256-EOhSmaRoTK9KOl1aipOh6pLWpZanH9rlqehhXoPflMQ="
+"sha256-Z2acBPmqrMKTbQ7mUOi3B8tIlCnLGJBAmu4X8Q5ufgg="